backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
cianparser
requests
tenacity
apscheduler